+++ src/DevilDaggersInfo.Tools.Engine/ImGui/ImGuiController.cs
@@ -34,43 +34,17 @@ public sealed class ImGuiController : IDisposable
 
 	private IntPtr _context;
 
-	/// <summary>
-	/// Constructs a new ImGuiController with font configuration.
-	/// </summary>
-	public ImGuiController(GL gl, IView view, IInputContext input, ImGuiFontConfig imGuiFontConfig)
-		: this(gl, view, input, imGuiFontConfig, null)
-	{
-	}
-
-	/// <summary>
-	/// Constructs a new ImGuiController with an onConfigureIO Action.
-	/// </summary>
 	public ImGuiController(GL gl, IView view, IInputContext input, Action onConfigureIo)
-		: this(gl, view, input, null, onConfigureIo)
-	{
-	}
-
-	/// <summary>
-	/// Constructs a new ImGuiController with font configuration and onConfigure Action.
-	/// </summary>
-	public ImGuiController(GL gl, IView view, IInputContext input, ImGuiFontConfig? imGuiFontConfig = null, Action? onConfigureIo = null)
 	{
 		Init(gl, view, input);
 
 		ImGuiIOPtr io = ImGuiNET.ImGui.GetIO();
-		if (imGuiFontConfig is not null)
-		{
-			IntPtr glyphRange = imGuiFontConfig.Value.GetGlyphRange?.Invoke(io) ?? default(IntPtr);
 
-			io.Fonts.AddFontFromFileTTF(imGuiFontConfig.Value.FontPath, imGuiFontConfig.Value.FontSize, null, glyphRange);
-		}
-
-		onConfigureIo?.Invoke();
+		onConfigureIo.Invoke();
 
 		io.BackendFlags |= ImGuiBackendFlags.RendererHasVtxOffset;
 
 		CreateDeviceResources();
-		SetKeyMappings();
 
 		SetPerFrameImGuiData(1f / 60f);
 
@@ -92,7 +66,6 @@ public sealed class ImGuiController : IDisposable
 
 	private void CreateDeviceResources()
 	{
-		// Backup GL state
 		_gl.GetInteger(GLEnum.TextureBinding2D, out int lastTexture);
 		_gl.GetInteger(GLEnum.ArrayBufferBinding, out int lastArrayBuffer);
 		_gl.GetInteger(GLEnum.VertexArrayBinding, out int lastVertexArray);
@@ -138,27 +111,20 @@ public sealed class ImGuiController : IDisposable
 		_vboHandle = _gl.GenBuffer();
 		_elementsHandle = _gl.GenBuffer();
 
-		RecreateFontDeviceTexture();
+		RecreateFontTexture();
 
-		// Restore modified GL state
 		_gl.BindTexture(GLEnum.Texture2D, (uint)lastTexture);
 		_gl.BindBuffer(GLEnum.ArrayBuffer, (uint)lastArrayBuffer);
 
 		_gl.BindVertexArray((uint)lastVertexArray);
 	}
 
-	/// <summary>
-	/// Creates the texture used to render text.
-	/// </summary>
-	private void RecreateFontDeviceTexture()
+	private void RecreateFontTexture()
 	{
-		// Build texture atlas
 		ImGuiIOPtr io = ImGuiNET.ImGui.GetIO();
 
-		// Load as RGBA 32-bit (75% of the memory is wasted, but default font is so small) because it is more likely to be compatible with user's existing shaders. If your ImTextureId represent a higher-level concept than just a GL texture id, consider calling GetTexDataAsAlpha8() instead to save on GPU memory.
 		io.Fonts.GetTexDataAsRGBA32(out IntPtr pixels, out int width, out int height, out int _);
 
-		// Upload texture to graphics system
 		_gl.GetInteger(GLEnum.TextureBinding2D, out int lastTexture);
 
 		_fontTexture = new Texture(_gl, width, height, pixels);
@@ -166,10 +132,8 @@ public sealed class ImGuiController : IDisposable
 		_fontTexture.SetMagFilter(TextureMagFilter.Linear);
 		_fontTexture.SetMinFilter(TextureMinFilter.Linear);
 
-		// Store our identifier
 		io.Fonts.SetTexID((IntPtr)_fontTexture.GlTexture);
 
-		// Restore state
 		_gl.BindTexture(GLEnum.Texture2D, (uint)lastTexture);
 	}
 
@@ -193,31 +157,25 @@ public sealed class ImGuiController : IDisposable
 		_windowHeight = size.Y;
 	}
 
-	/// <summary>
-	/// Renders the ImGui draw list data.
-	/// This method requires a <see cref="GraphicsDevice"/> because it may create new DeviceBuffers if the size of vertex
-	/// or index data has increased beyond the capacity of the existing buffers.
-	/// A <see cref="CommandList"/> is needed to submit drawing and resource update commands.
-	/// </summary>
 	public void Render()
 	{
-		if (_frameBegun)
+		if (!_frameBegun)
+			return;
+
+		IntPtr oldCtx = ImGuiNET.ImGui.GetCurrentContext();
+
+		if (oldCtx != _context)
 		{
-			IntPtr oldCtx = ImGuiNET.ImGui.GetCurrentContext();
+			ImGuiNET.ImGui.SetCurrentContext(_context);
+		}
 
-			if (oldCtx != _context)
-			{
-				ImGuiNET.ImGui.SetCurrentContext(_context);
-			}
+		_frameBegun = false;
+		ImGuiNET.ImGui.Render();
+		RenderImDrawData(ImGuiNET.ImGui.GetDrawData());
 
-			_frameBegun = false;
-			ImGuiNET.ImGui.Render();
-			RenderImDrawData(ImGuiNET.ImGui.GetDrawData());
-
-			if (oldCtx != _context)
-			{
-				ImGuiNET.ImGui.SetCurrentContext(oldCtx);
-			}
+		if (oldCtx != _context)
+		{
+			ImGuiNET.ImGui.SetCurrentContext(oldCtx);
 		}
 	}
 
@@ -248,6 +206,14 @@ public sealed class ImGuiController : IDisposable
 		{
 			ImGuiNET.ImGui.SetCurrentContext(oldCtx);
 		}
+
+		// Since ImGui.NET 1.90.0.1, ImGui.NewFrame() overrides the key modifiers, so we need to set them after instead of before.
+		ImGuiIOPtr io = ImGuiNET.ImGui.GetIO();
+		IKeyboard keyboardState = _input.Keyboards[0];
+		io.KeyCtrl = keyboardState.IsKeyPressed(Key.ControlLeft) || keyboardState.IsKeyPressed(Key.ControlRight);
+		io.KeyAlt = keyboardState.IsKeyPressed(Key.AltLeft) || keyboardState.IsKeyPressed(Key.AltRight);
+		io.KeyShift = keyboardState.IsKeyPressed(Key.ShiftLeft) || keyboardState.IsKeyPressed(Key.ShiftRight);
+		io.KeySuper = keyboardState.IsKeyPressed(Key.SuperLeft) || keyboardState.IsKeyPressed(Key.SuperRight);
 	}
 
 	/// <summary>
@@ -257,7 +223,7 @@ public sealed class ImGuiController : IDisposable
 	private void SetPerFrameImGuiData(float deltaSeconds)
 	{
 		ImGuiIOPtr io = ImGuiNET.ImGui.GetIO();
-		io.DisplaySize = new Vector2(_windowWidth, _windowHeight);
+		io.DisplaySize = new(_windowWidth, _windowHeight);
 
 		if (_windowWidth > 0 && _windowHeight > 0)
 			io.DisplayFramebufferScale = new Vector2(_view.FramebufferSize.X / _windowWidth, _view.FramebufferSize.Y / _windowHeight);
@@ -287,7 +253,7 @@ public sealed class ImGuiController : IDisposable
 			if (key == Key.Unknown)
 				continue;
 
-			io.KeysDown[(int)key] = keyboardState.IsKeyPressed(key);
+			io.AddKeyEvent(GetImGuiKey(key), keyboardState.IsKeyPressed(key));
 		}
 
 		for (int i = 0; i < _pressedChars.Count; i++)
@@ -297,36 +263,6 @@ public sealed class ImGuiController : IDisposable
 		}
 
 		_pressedChars.Clear();
-
-		io.KeyCtrl = keyboardState.IsKeyPressed(Key.ControlLeft) || keyboardState.IsKeyPressed(Key.ControlRight);
-		io.KeyAlt = keyboardState.IsKeyPressed(Key.AltLeft) || keyboardState.IsKeyPressed(Key.AltRight);
-		io.KeyShift = keyboardState.IsKeyPressed(Key.ShiftLeft) || keyboardState.IsKeyPressed(Key.ShiftRight);
-		io.KeySuper = keyboardState.IsKeyPressed(Key.SuperLeft) || keyboardState.IsKeyPressed(Key.SuperRight);
-	}
-
-	private static void SetKeyMappings()
-	{
-		ImGuiIOPtr io = ImGuiNET.ImGui.GetIO();
-		io.KeyMap[(int)ImGuiKey.Tab] = (int)Key.Tab;
-		io.KeyMap[(int)ImGuiKey.LeftArrow] = (int)Key.Left;
-		io.KeyMap[(int)ImGuiKey.RightArrow] = (int)Key.Right;
-		io.KeyMap[(int)ImGuiKey.UpArrow] = (int)Key.Up;
-		io.KeyMap[(int)ImGuiKey.DownArrow] = (int)Key.Down;
-		io.KeyMap[(int)ImGuiKey.PageUp] = (int)Key.PageUp;
-		io.KeyMap[(int)ImGuiKey.PageDown] = (int)Key.PageDown;
-		io.KeyMap[(int)ImGuiKey.Home] = (int)Key.Home;
-		io.KeyMap[(int)ImGuiKey.End] = (int)Key.End;
-		io.KeyMap[(int)ImGuiKey.Delete] = (int)Key.Delete;
-		io.KeyMap[(int)ImGuiKey.Backspace] = (int)Key.Backspace;
-		io.KeyMap[(int)ImGuiKey.Enter] = (int)Key.Enter;
-		io.KeyMap[(int)ImGuiKey.KeypadEnter] = (int)Key.KeypadEnter;
-		io.KeyMap[(int)ImGuiKey.Escape] = (int)Key.Escape;
-		io.KeyMap[(int)ImGuiKey.A] = (int)Key.A;
-		io.KeyMap[(int)ImGuiKey.C] = (int)Key.C;
-		io.KeyMap[(int)ImGuiKey.V] = (int)Key.V;
-		io.KeyMap[(int)ImGuiKey.X] = (int)Key.X;
-		io.KeyMap[(int)ImGuiKey.Y] = (int)Key.Y;
-		io.KeyMap[(int)ImGuiKey.Z] = (int)Key.Z;
 	}
 
 	private unsafe void SetupRenderState(ImDrawDataPtr drawDataPtr)
@@ -347,7 +283,7 @@ public sealed class ImGuiController : IDisposable
 		float t = drawDataPtr.DisplayPos.Y;
 		float b = drawDataPtr.DisplayPos.Y + drawDataPtr.DisplaySize.Y;
 
-		Span<float> orthoProjection = stackalloc float[]
+		Span<float> orthographicProjection = stackalloc float[]
 		{
 			2.0f / (r - l), 0.0f, 0.0f, 0.0f,
 			0.0f, 2.0f / (t - b), 0.0f, 0.0f,
@@ -357,7 +293,7 @@ public sealed class ImGuiController : IDisposable
 
 		_shader.UseShader();
 		_gl.Uniform1(_attribLocationTex, 0);
-		_gl.UniformMatrix4(_attribLocationProjMtx, 1, false, orthoProjection);
+		_gl.UniformMatrix4(_attribLocationProjMtx, 1, false, orthographicProjection);
 
 		_gl.BindSampler(0, 0);
 
@@ -447,7 +383,7 @@ public sealed class ImGuiController : IDisposable
 				clipRect.Z = (cmdPtr.ClipRect.Z - clipOff.X) * clipScale.X;
 				clipRect.W = (cmdPtr.ClipRect.W - clipOff.Y) * clipScale.Y;
 
-				if (clipRect.X < framebufferWidth && clipRect.Y < framebufferHeight && clipRect.Z >= 0.0f && clipRect.W >= 0.0f)
+				if (clipRect.X < framebufferWidth && clipRect.Y < framebufferHeight && clipRect is { Z: >= 0.0f, W: >= 0.0f })
 				{
 					// Apply scissor/clipping rectangle
 					_gl.Scissor((int)clipRect.X, (int)(framebufferHeight - clipRect.W), (uint)(clipRect.Z - clipRect.X), (uint)(clipRect.W - clipRect.Y));
@@ -529,5 +465,67 @@ public sealed class ImGuiController : IDisposable
 		_shader.Dispose();
 
 		ImGuiNET.ImGui.DestroyContext(_context);
+	}
+
+	private static ImGuiKey GetImGuiKey(Key key)
+	{
+		return key switch
+		{
+			>= Key.Number0 and <= Key.Number9 => key - Key.D0 + ImGuiKey._0,
+			>= Key.A and <= Key.Z => key - Key.A + ImGuiKey.A,
+			>= Key.Keypad0 and <= Key.Keypad9 => key - Key.Keypad0 + ImGuiKey.Keypad0,
+			>= Key.F1 and <= Key.F24 => key - Key.F1 + ImGuiKey.F24,
+			_ => key switch
+			{
+				Key.Tab => ImGuiKey.Tab,
+				Key.Left => ImGuiKey.LeftArrow,
+				Key.Right => ImGuiKey.RightArrow,
+				Key.Up => ImGuiKey.UpArrow,
+				Key.Down => ImGuiKey.DownArrow,
+				Key.PageUp => ImGuiKey.PageUp,
+				Key.PageDown => ImGuiKey.PageDown,
+				Key.Home => ImGuiKey.Home,
+				Key.End => ImGuiKey.End,
+				Key.Insert => ImGuiKey.Insert,
+				Key.Delete => ImGuiKey.Delete,
+				Key.Backspace => ImGuiKey.Backspace,
+				Key.Space => ImGuiKey.Space,
+				Key.Enter => ImGuiKey.Enter,
+				Key.Escape => ImGuiKey.Escape,
+				Key.Apostrophe => ImGuiKey.Apostrophe,
+				Key.Comma => ImGuiKey.Comma,
+				Key.Minus => ImGuiKey.Minus,
+				Key.Period => ImGuiKey.Period,
+				Key.Slash => ImGuiKey.Slash,
+				Key.Semicolon => ImGuiKey.Semicolon,
+				Key.Equal => ImGuiKey.Equal,
+				Key.LeftBracket => ImGuiKey.LeftBracket,
+				Key.BackSlash => ImGuiKey.Backslash,
+				Key.RightBracket => ImGuiKey.RightBracket,
+				Key.GraveAccent => ImGuiKey.GraveAccent,
+				Key.CapsLock => ImGuiKey.CapsLock,
+				Key.ScrollLock => ImGuiKey.ScrollLock,
+				Key.NumLock => ImGuiKey.NumLock,
+				Key.PrintScreen => ImGuiKey.PrintScreen,
+				Key.Pause => ImGuiKey.Pause,
+				Key.KeypadDecimal => ImGuiKey.KeypadDecimal,
+				Key.KeypadDivide => ImGuiKey.KeypadDivide,
+				Key.KeypadMultiply => ImGuiKey.KeypadMultiply,
+				Key.KeypadSubtract => ImGuiKey.KeypadSubtract,
+				Key.KeypadAdd => ImGuiKey.KeypadAdd,
+				Key.KeypadEnter => ImGuiKey.KeypadEnter,
+				Key.KeypadEqual => ImGuiKey.KeypadEqual,
+				Key.ShiftLeft => ImGuiKey.LeftShift,
+				Key.ControlLeft => ImGuiKey.LeftCtrl,
+				Key.AltLeft => ImGuiKey.LeftAlt,
+				Key.SuperLeft => ImGuiKey.LeftSuper,
+				Key.ShiftRight => ImGuiKey.RightShift,
+				Key.ControlRight => ImGuiKey.RightCtrl,
+				Key.AltRight => ImGuiKey.RightAlt,
+				Key.SuperRight => ImGuiKey.RightSuper,
+				Key.Menu => ImGuiKey.Menu,
+				_ => ImGuiKey.None,
+			},
+		};
 	}
 }